+++ planck-cljs/test/planck/repl_test.cljs
@@ -11,16 +11,16 @@
 
 (deftest get-highlight-coords
   (testing "Highlight coordinates"
-    (is (= [0 0] (js->clj (planck.repl/get-highlight-coords 1 "[]" []))))
-    (is (= [0 1] (js->clj (planck.repl/get-highlight-coords 2 "[[]]" []))))
-    (is (= [0 0] (js->clj (planck.repl/get-highlight-coords 3 "[()]" []))))
-    (is (= [0 1] (js->clj (planck.repl/get-highlight-coords 2 " []" []))))
-    (is (= [1 0] (js->clj (planck.repl/get-highlight-coords 1 "]" ["["]))))
-    (is (= [2 0] (js->clj (planck.repl/get-highlight-coords 1 "]" ["[" ""]))))
-    (is (= [1 0] (js->clj (planck.repl/get-highlight-coords 1 "]" ["[" "["]))))
-    (is (= [2 0] (js->clj (planck.repl/get-highlight-coords 1 "]" ["[" "[]"]))))
-    (is (= [0 1] (js->clj (planck.repl/get-highlight-coords 2 "#{}" []))))
-    (is (= [0 0] (js->clj (planck.repl/get-highlight-coords 4 "[\"[\"]" []))))))
+    (is (= [0 0] (js->clj (#'planck.repl/get-highlight-coords 1 "[]" []))))
+    (is (= [0 1] (js->clj (#'planck.repl/get-highlight-coords 2 "[[]]" []))))
+    (is (= [0 0] (js->clj (#'planck.repl/get-highlight-coords 3 "[()]" []))))
+    (is (= [0 1] (js->clj (#'planck.repl/get-highlight-coords 2 " []" []))))
+    (is (= [1 0] (js->clj (#'planck.repl/get-highlight-coords 1 "]" ["["]))))
+    (is (= [2 0] (js->clj (#'planck.repl/get-highlight-coords 1 "]" ["[" ""]))))
+    (is (= [1 0] (js->clj (#'planck.repl/get-highlight-coords 1 "]" ["[" "["]))))
+    (is (= [2 0] (js->clj (#'planck.repl/get-highlight-coords 1 "]" ["[" "[]"]))))
+    (is (= [0 1] (js->clj (#'planck.repl/get-highlight-coords 2 "#{}" []))))
+    (is (= [0 0] (js->clj (#'planck.repl/get-highlight-coords 4 "[\"[\"]" []))))))
 
 (deftest test-apropos
   (is (= '(cljs.core/ffirst) (planck.repl/apropos "ffirst")))
@@ -33,62 +33,62 @@
 
 (deftest get-error-indicator-test
   (is (= "             ^"
-        (repl/get-error-column-indicator
-          (ex-info repl/could-not-eval-expr {}
+        (#'repl/get-error-column-indicator
+          (ex-info @#'repl/could-not-eval-expr {}
             (ex-info "" {:tag    :cljs/analysis-error
                          :column 3}))
           "foo.core"))))
 
 (deftest undo-reader-conditional-whitespace-docstring-test
-  (is (= "a\n  b" (planck.repl/undo-reader-conditional-whitespace-docstring "a\n     b"))))
+  (is (= "a\n  b" (#'planck.repl/undo-reader-conditional-whitespace-docstring "a\n     b"))))
 
 (deftest root-resource-test
-  (is (= "/foo_bar_baz/boo/core" (planck.repl/root-resource 'foo-bar-baz.boo.core))))
+  (is (= "/foo_bar_baz/boo/core" (#'planck.repl/root-resource 'foo-bar-baz.boo.core))))
 
 (defspec add-drop-macros-suffix-test
   (prop/for-all [ns-sym gen/symbol]
     (= ns-sym (-> ns-sym
-                repl/add-macros-suffix
+                ^:private-var-access-nowarn repl/add-macros-suffix
                 str
-                repl/drop-macros-suffix
+                ^:private-var-access-nowarn repl/drop-macros-suffix
                 symbol))))
 
 (deftest get-arglists-test
-  (is (= '([x] [x y] [x y & more]) (planck.repl/get-arglists "max")))
-  (is (nil? (planck.repl/get-arglists "bogus-undefined"))))
+  (is (= '([x] [x y] [x y & more]) (#'planck.repl/get-arglists "max")))
+  (is (nil? (#'planck.repl/get-arglists "bogus-undefined"))))
 
 (deftest completions-for-goog-ns
-  (is (some #{"isArrayLike"} (planck.repl/completion-candidates-for-ns 'goog false)))
-  (is (some #{"trimLeft"} (planck.repl/completion-candidates-for-ns 'goog.string false))))
+  (is (some #{"isArrayLike"} (#'planck.repl/completion-candidates-for-ns 'goog false)))
+  (is (some #{"trimLeft"} (#'planck.repl/completion-candidates-for-ns 'goog.string false))))
 
 (deftest doc-test
   (is (empty? (with-out-str (planck.repl/doc every)))))
 
 (deftest planck-native?-test
-  (is (planck.repl/planck-native? 'PLANCK_SOCKET_CLOSE))
-  (is (not (planck.repl/planck-native? 'foobar))))
+  (is (#'planck.repl/planck-native? 'PLANCK_SOCKET_CLOSE))
+  (is (not (#'planck.repl/planck-native? 'foobar))))
 
 (deftest gensym?-test
-  (is (planck.repl/gensym? (gensym)))
-  (is (not (planck.repl/gensym? 'foobar))))
+  (is (#'planck.repl/gensym? (gensym)))
+  (is (not (#'planck.repl/gensym? 'foobar))))
 
 (deftest test-is-readable?
-  (is (= (planck.repl/is-readable? "(") nil))
-  (is (= (planck.repl/is-readable? "(+ 1 2)") ""))
-  (is (= (planck.repl/is-readable? "(+ 1 2) :foo") " :foo"))
-  (is (= (planck.repl/is-readable? "") nil))
-  (is (= (planck.repl/is-readable? ")") "")))
+  (is (= (#'planck.repl/is-readable? "(") nil))
+  (is (= (#'planck.repl/is-readable? "(+ 1 2)") ""))
+  (is (= (#'planck.repl/is-readable? "(+ 1 2) :foo") " :foo"))
+  (is (= (#'planck.repl/is-readable? "") nil))
+  (is (= (#'planck.repl/is-readable? ")") "")))
 
 (deftest repl-read-string-test
-  (is (thrown? js/Error (planck.repl/repl-read-string "34f")))
-  (is (thrown? js/Error (planck.repl/repl-read-string "a:")))
-  (is (thrown? js/Error (planck.repl/repl-read-string "]"))))
+  (is (thrown? js/Error (#'planck.repl/repl-read-string "34f")))
+  (is (thrown? js/Error (#'planck.repl/repl-read-string "a:")))
+  (is (thrown? js/Error (#'planck.repl/repl-read-string "]"))))
 
 (deftest strip-source-map-test
   (let [input-sm {0 {2 [{:line 1 :col 1} {:line 2 :col 2 :name "a"}]
                      3 [{:line 2 :col 2 :name "b"}]}}
         stripped {0 {2 [{:line 2 :col 2}]}}]
-    (is (= stripped (planck.repl/strip-source-map input-sm)))))
+    (is (= stripped (#'planck.repl/strip-source-map input-sm)))))
 
 (deftest require-goog-test
   (is (false? (g/isArrayLike nil)))